Exercise 4.21: Bloglist expansion step 9
Exercise 4.21: Bloglist exapnsion step 9. The application is modified so that a blog can only be deleted if the token sent with the request is the same as that of the blog's creator.

+++ blogposts/controllers/blogs.js
@@ -20,11 +20,12 @@ blogsRouter.post('/', async (request, response) => {
     const body = request.body
     const token = getTokenFrom(request)
     const decodedToken = jwt.verify(token, process.env.SECRET)
+    //    console.log(decodedToken)
     if (!token || !decodedToken.id) {
         return response.status(401).json({ error: 'token missing or invalid' })
     }
     const user = await User.findById(decodedToken.id)
-    
+
     const blog = new Blog({
         title: body.title,
         author: body.author,
@@ -41,12 +42,22 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-
-    const blogToDelete = await Blog.findByIdAndRemove(request.params.id)
-    if (blogToDelete) {
-        response.status(204).end()
+    const token = getTokenFrom(request)
+    const decodedToken = jwt.verify(token, process.env.SECRET)
+    const userid = decodedToken.id
+    if (!token || !decodedToken.id) {
+        return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    const blog = await Blog.findById(request.params.id)
+    if (blog.user.toString() === userid.toString()) {
+        const blogToDelete = await Blog.findByIdAndRemove(request.params.id)
+        if (blogToDelete) {
+            response.status(204).end()
+        } else {
+            response.status(404).send(`The data is already deleted from database`)
+        }
     } else {
-        response.status(404).send(`The data is already deleted from database`)
+        response.status(404).send('Permission to delete the data is denied')
     }
 })
